+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Lim 2025 amikacin ([doi:10.1128/aac.00818-25](https://doi.org/10.1128/aac.00818-25)) -- preterm and term neonates receiving intravenous or intramuscular amikacin.
+- Add Calderin 2025 pyrazinamide and isoniazid plasma/CSF models ([doi:10.1128/aac.00099-25](https://doi.org/10.1128/aac.00099-25)) -- South African adults with HIV-associated tuberculous meningitis.
 
 - Fix a dropped `exp()` in the bacteremia term of
   `Nagy_2017_obiltoxaximab_survival`. `logit(psurv)` now uses